get_column: force strings to be strings (stop checking type)
SQLITE3 appears to ignore the column type in a CREATE TABLE!  if you
INSERT a string to a table that contains all decimal digits, SQLITE3
does something very strange: it stores it as SQLITE_INT! this means
this validation fails. but if you call sqlite_column_text, it will
convert it back to a string for you. so, disable the type check in the
get_column_string method to coerce strings to be strings.